Move dispatcher to separate file
